requirements.txt
scrapy>=2.8
python-dateutil>=2.8
tzdata>=2023.3
itemadapter>=0.8.0
